Fix: Open activity does not work from lock screen notification
This changes the PendingIntent for the Open action button to
ACTION_VIEW, thus bypassing the IntentService. The activity was only
started behind the lock screen without unlocking the device in the
previous setup.

As a side effect, pressing the Open action does not dismiss the
notification.

The issue that Chrome does not load the URL when the activity is
initiated from lock screen is migitated by adding extra data to
instruct Chrome to create a custom tab. Other browsers ignore this
data.

FPIIM-1453

Change-Id: I2f68430ac498a56df70fbf3170f1f334d4041add