Delegate display of instructions to CameraSwapInfo
ModuleDetect needs to be privileged to be able to read sysfs files for
detection of camera modules. WebView cannot be used from a privileged
context due to security considerations.

This change separates module detection and notification on the one hand
from actually displaying information in an activity on the other hand.
The latter is handled by CameraSwapInfo.

FPIIM-1467

Change-Id: Idaeccca592fad85b22a2d3bd44616882351064ec